// frontend/src/types/proto/api/v1alpha/issue_service.ts
export enum IssueStatus {
  ISSUE_STATUS_UNSPECIFIED = "ISSUE_STATUS_UNSPECIFIED",
  OPEN = "OPEN",
  DONE = "DONE",
  CANCELED = "CANCELED",
  UNRECOGNIZED = "UNRECOGNIZED",
}
// ...
export interface ListIssueCommentsRequest {
  /** Format: projects/{projects}/issues/{issue} */
  parent: string;
  /**
   * The maximum number of issue comments to return. The service may return fewer than
   * this value.
   * If unspecified, at most 10 issue comments will be returned.
   * The maximum value is 1000; values above 1000 will be coerced to 1000.
   */
  pageSize: number;
  /**
   * A page token, received from a previous `ListIssueComments` call.
   * Provide this to retrieve the subsequent page.
   *
   * When paginating, all other parameters provided to `ListIssueComments` must match
   * the call that provided the page token.
   */
  pageToken: string;
}
// ...
export interface IssueComment {
  /** Format: projects/{project}/issues/{issue}/issueComments/{issueComment-uid} */
  name: string;
  comment: string;
  /** TODO: use struct message instead. */
  payload: string;
  createTime: Timestamp | undefined;
  updateTime:
    | Timestamp
    | undefined;
  /** Format: users/{email} */
  creator: string;
  approval?: IssueComment_Approval | undefined;
  issueUpdate?: IssueComment_IssueUpdate | undefined;
  stageEnd?: IssueComment_StageEnd | undefined;
  taskUpdate?: IssueComment_TaskUpdate | undefined;
  taskPriorBackup?: IssueComment_TaskPriorBackup | undefined;
}

export interface IssueComment_Approval {
  status: IssueComment_Approval_Status;
}

export enum IssueComment_Approval_Status {
  STATUS_UNSPECIFIED = "STATUS_UNSPECIFIED",
  PENDING = "PENDING",
  APPROVED = "APPROVED",
  REJECTED = "REJECTED",
  UNRECOGNIZED = "UNRECOGNIZED",
}
// ...
export interface IssueComment_IssueUpdate {
  fromTitle?: string | undefined;
  toTitle?: string | undefined;
  fromDescription?: string | undefined;
  toDescription?: string | undefined;
  fromStatus?: IssueStatus | undefined;
  toStatus?: IssueStatus | undefined;
  fromLabels: string[];
  toLabels: string[];
}

export interface IssueComment_StageEnd {
  stage: string;
}

export interface IssueComment_TaskUpdate {
  tasks: string[];
  /** Format: projects/{project}/sheets/{sheet} */
  fromSheet?:
    | string
    | undefined;
  /** Format: projects/{project}/sheets/{sheet} */
  toSheet?: string | undefined;
  fromEarliestAllowedTime?: Timestamp | undefined;
  toEarliestAllowedTime?: Timestamp | undefined;
  toStatus?: IssueComment_TaskUpdate_Status | undefined;
}

export enum IssueComment_TaskUpdate_Status {
  STATUS_UNSPECIFIED = "STATUS_UNSPECIFIED",
  PENDING = "PENDING",
  RUNNING = "RUNNING",
  DONE = "DONE",
  FAILED = "FAILED",
  SKIPPED = "SKIPPED",
  CANCELED = "CANCELED",
  UNRECOGNIZED = "UNRECOGNIZED",
}
// ...
export interface IssueComment_TaskPriorBackup {
  task: string;
  tables: IssueComment_TaskPriorBackup_Table[];
  originalLine?: number | undefined;
  database: string;
  error: string;
}

export interface IssueComment_TaskPriorBackup_Table {
  schema: string;
  table: string;
}
// ...
export const ListIssueCommentsRequest: MessageFns<ListIssueCommentsRequest> = {
  encode(message: ListIssueCommentsRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.parent !== "") {
      writer.uint32(10).string(message.parent);
    }
    if (message.pageSize !== 0) {
      writer.uint32(16).int32(message.pageSize);
    }
    if (message.pageToken !== "") {
      writer.uint32(26).string(message.pageToken);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ListIssueCommentsRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseListIssueCommentsRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.parent = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.pageSize = reader.int32();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.pageToken = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): ListIssueCommentsRequest {
    return {
      parent: isSet(object.parent) ? globalThis.String(object.parent) : "",
      pageSize: isSet(object.pageSize) ? globalThis.Number(object.pageSize) : 0,
      pageToken: isSet(object.pageToken) ? globalThis.String(object.pageToken) : "",
    };
  },

  toJSON(message: ListIssueCommentsRequest): unknown {
    const obj: any = {};
    if (message.parent !== "") {
      obj.parent = message.parent;
    }
    if (message.pageSize !== 0) {
      obj.pageSize = Math.round(message.pageSize);
    }
    if (message.pageToken !== "") {
      obj.pageToken = message.pageToken;
    }
    return obj;
  },

  create(base?: DeepPartial<ListIssueCommentsRequest>): ListIssueCommentsRequest {
    return ListIssueCommentsRequest.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<ListIssueCommentsRequest>): ListIssueCommentsRequest {
    const message = createBaseListIssueCommentsRequest();
    message.parent = object.parent ?? "";
    message.pageSize = object.pageSize ?? 0;
    message.pageToken = object.pageToken ?? "";
    return message;
  },
};
// ...
export const IssueComment: MessageFns<IssueComment> = {
// ...
  fromJSON(object: any): IssueComment {
    return {
      name: isSet(object.name) ? globalThis.String(object.name) : "",
      comment: isSet(object.comment) ? globalThis.String(object.comment) : "",
      payload: isSet(object.payload) ? globalThis.String(object.payload) : "",
      createTime: isSet(object.createTime) ? fromJsonTimestamp(object.createTime) : undefined,
      updateTime: isSet(object.updateTime) ? fromJsonTimestamp(object.updateTime) : undefined,
      creator: isSet(object.creator) ? globalThis.String(object.creator) : "",
      approval: isSet(object.approval) ? IssueComment_Approval.fromJSON(object.approval) : undefined,
      issueUpdate: isSet(object.issueUpdate) ? IssueComment_IssueUpdate.fromJSON(object.issueUpdate) : undefined,
      stageEnd: isSet(object.stageEnd) ? IssueComment_StageEnd.fromJSON(object.stageEnd) : undefined,
      taskUpdate: isSet(object.taskUpdate) ? IssueComment_TaskUpdate.fromJSON(object.taskUpdate) : undefined,
      taskPriorBackup: isSet(object.taskPriorBackup)
        ? IssueComment_TaskPriorBackup.fromJSON(object.taskPriorBackup)
        : undefined,
    };
  },
// ...
};

function createBaseIssueComment_Approval(): IssueComment_Approval {
  return { status: IssueComment_Approval_Status.STATUS_UNSPECIFIED };
}

export const IssueComment_Approval: MessageFns<IssueComment_Approval> = {
  encode(message: IssueComment_Approval, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.status !== IssueComment_Approval_Status.STATUS_UNSPECIFIED) {
      writer.uint32(8).int32(issueComment_Approval_StatusToNumber(message.status));
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): IssueComment_Approval {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseIssueComment_Approval();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.status = issueComment_Approval_StatusFromJSON(reader.int32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): IssueComment_Approval {
    return {
      status: isSet(object.status)
        ? issueComment_Approval_StatusFromJSON(object.status)
        : IssueComment_Approval_Status.STATUS_UNSPECIFIED,
    };
  },

  toJSON(message: IssueComment_Approval): unknown {
    const obj: any = {};
    if (message.status !== IssueComment_Approval_Status.STATUS_UNSPECIFIED) {
      obj.status = issueComment_Approval_StatusToJSON(message.status);
    }
    return obj;
  },

  create(base?: DeepPartial<IssueComment_Approval>): IssueComment_Approval {
    return IssueComment_Approval.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<IssueComment_Approval>): IssueComment_Approval {
    const message = createBaseIssueComment_Approval();
    message.status = object.status ?? IssueComment_Approval_Status.STATUS_UNSPECIFIED;
    return message;
  },
};
// ...
export const IssueComment_StageEnd: MessageFns<IssueComment_StageEnd> = {
  encode(message: IssueComment_StageEnd, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.stage !== "") {
      writer.uint32(10).string(message.stage);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): IssueComment_StageEnd {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseIssueComment_StageEnd();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.stage = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): IssueComment_StageEnd {
    return { stage: isSet(object.stage) ? globalThis.String(object.stage) : "" };
  },

  toJSON(message: IssueComment_StageEnd): unknown {
    const obj: any = {};
    if (message.stage !== "") {
      obj.stage = message.stage;
    }
    return obj;
  },

  create(base?: DeepPartial<IssueComment_StageEnd>): IssueComment_StageEnd {
    return IssueComment_StageEnd.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<IssueComment_StageEnd>): IssueComment_StageEnd {
    const message = createBaseIssueComment_StageEnd();
    message.stage = object.stage ?? "";
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends Long ? string | number | Long : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;
// ...
export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create(base?: DeepPartial<T>): T;
  fromPartial(object: DeepPartial<T>): T;
}